pyproject.toml
[project]
name = "automaton-auditor"
version = "0.1.0"
description = "Production-grade AI multi-agent audit system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "docling>=2.0",
    "typing-extensions>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.4",
]
